requirements.txt
ultralytics
opencv-python
pandas
numpy
matplotlib
seaborn
pyyaml
tqdm
Pillow
streamlit
easyocr
moviepy